+++ crates/ruma-client-api/CHANGELOG.md
@@ -18,6 +18,7 @@ Improvements:
 Bug fixes:
 
 - Rename `avatar` to `avatar_url` when (De)serializing `SlidingSyncRoomHero`
+- Make `id_server` and `id_access_token` in `ThirdpartyIdCredentials` optional fields
 - `user_id` of `SlidingSyncRoomHero` is now mandatory
 - Make authentication with access token optional for the `change_password` and
   `deactivate` endpoints.